fix: move logic from app to spinner component
