+++ src/components/LoaderSpinner.vue
@@ -1,12 +1,14 @@
 <template>
-  <div class="loader-outer">
+  <div class="loader-outer" v-show="isLoading">
     <ProgressSpinner />
   </div>
 </template>
 
 <script lang="ts">
-import { defineComponent } from 'vue';
+import { defineComponent, ref, computed } from 'vue';
 import ProgressSpinner from 'primevue/progressspinner';
+import { HTTP, HTTP_AUTH } from '@/core/api/http-common';
+import { AxiosError, AxiosRequestConfig, AxiosResponse } from 'axios';
 
 export default defineComponent({
   name: 'LoaderSpinner',
@@ -14,7 +16,64 @@ export default defineComponent({
     ProgressSpinner,
   },
   setup() {
-    return {};
+    const isLoading = ref(false);
+
+    computed(() => {
+      return isLoading.value;
+    });
+
+    const showLoader = () => {
+      isLoading.value = true;
+    };
+    const hideLoader = () => {
+      isLoading.value = false;
+    };
+
+    HTTP.interceptors.request.use(
+      (req: AxiosRequestConfig) => {
+        showLoader();
+        return req;
+      },
+      (error: AxiosError) => {
+        hideLoader();
+        return error;
+      }
+    );
+
+    HTTP.interceptors.response.use(
+      (response: AxiosResponse) => {
+        hideLoader();
+        return response;
+      },
+      (error: AxiosError) => {
+        hideLoader();
+        return error;
+      }
+    );
+
+    HTTP_AUTH.interceptors.request.use(
+      (req: AxiosRequestConfig) => {
+        showLoader();
+        return req;
+      },
+      (error: AxiosError) => {
+        hideLoader();
+        return error;
+      }
+    );
+
+    HTTP_AUTH.interceptors.response.use(
+      (response: AxiosResponse) => {
+        hideLoader();
+        return response;
+      },
+      (error: AxiosError) => {
+        hideLoader();
+        return error;
+      }
+    );
+
+    return { showLoader, hideLoader, isLoading };
   },
 });
 </script>
